Refactored/Reorganized IQueryable FilterBy extension methods.
