+++ Modix.Data/Models/Moderation/ModerationActionSearchCriteria.cs
@@ -1,6 +1,8 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 
 using Modix.Data.Repositories;
+using Modix.Data.Utilities;
 
 namespace Modix.Data.Models.Moderation
 {
@@ -28,5 +30,31 @@ namespace Modix.Data.Models.Moderation
         /// A <see cref="ModerationActionEntity.CreatedById"/> value, defining the <see cref="ModerationActionEntity"/> entities to be returned.
         /// </summary>
         public ulong? CreatedById { get; set; }
+    }
+
+    internal static class ModerationActionQueryableExtensions
+    {
+        public static IQueryable<ModerationActionEntity> FilterModerationActionsBy(this IQueryable<ModerationActionEntity> query, ModerationActionSearchCriteria criteria)
+        {
+            var longGuildId = (long?)criteria?.GuildId;
+            var longCreatedById = (long?)criteria?.CreatedById;
+
+            return query
+                .FilterBy(
+                    x => x.GuildId == longGuildId,
+                    longGuildId != null)
+                .FilterBy(
+                    x => criteria.Types.Contains(x.Type),
+                    criteria?.Types?.Any() ?? false)
+                .FilterBy(
+                    x => x.Created >= criteria.CreatedRange.Value.From,
+                    criteria?.CreatedRange?.From != null)
+                .FilterBy(
+                    x => x.Created <= criteria.CreatedRange.Value.To,
+                    criteria?.CreatedRange?.To != null)
+                .FilterBy(
+                    x => x.CreatedById == longCreatedById,
+                    longCreatedById != null);
+        }
     }
 }
