+++ Modix.Data/Repositories/InfractionRepository.cs
@@ -160,13 +160,13 @@ namespace Modix.Data.Repositories
         /// <inheritdoc />
         public Task<bool> AnyAsync(InfractionSearchCriteria criteria)
             => ModixContext.Infractions.AsNoTracking()
-                .FilterInfractionsBy(criteria)
+                .FilterBy(criteria)
                 .AnyAsync();
 
         /// <inheritdoc />
         public Task<DateTimeOffset?> ReadExpiresFirstOrDefaultAsync(InfractionSearchCriteria searchCriteria, IEnumerable<SortingCriteria> sortingCriteria = null)
             => ModixContext.Infractions.AsNoTracking()
-                .FilterInfractionsBy(searchCriteria)
+                .FilterBy(searchCriteria)
                 .Select(InfractionSummary.FromEntityProjection)
                 .SortBy(sortingCriteria, InfractionSummary.SortablePropertyMap)
                 .Select(x => x.Expires)
@@ -175,14 +175,14 @@ namespace Modix.Data.Repositories
         /// <inheritdoc />
         public async Task<IReadOnlyCollection<long>> SearchIdsAsync(InfractionSearchCriteria searchCriteria)
             => await ModixContext.Infractions.AsNoTracking()
-                .FilterInfractionsBy(searchCriteria)
+                .FilterBy(searchCriteria)
                 .Select(x => x.Id)
                 .ToArrayAsync();
 
         /// <inheritdoc />
         public async Task<IReadOnlyCollection<InfractionSummary>> SearchSummariesAsync(InfractionSearchCriteria searchCriteria, IEnumerable<SortingCriteria> sortingCriteria = null)
             => await ModixContext.Infractions.AsNoTracking()
-                .FilterInfractionsBy(searchCriteria)
+                .FilterBy(searchCriteria)
                 .Select(InfractionSummary.FromEntityProjection)
                 .SortBy(sortingCriteria, InfractionSummary.SortablePropertyMap)
                 .ToArrayAsync();
@@ -193,7 +193,7 @@ namespace Modix.Data.Repositories
             var sourceQuery = ModixContext.Infractions.AsNoTracking();
 
             var filteredQuery = sourceQuery
-                .FilterInfractionsBy(searchCriteria);
+                .FilterBy(searchCriteria);
 
             var pagedQuery = filteredQuery
                 .Select(InfractionSummary.FromEntityProjection)
@@ -278,47 +278,5 @@ namespace Modix.Data.Repositories
 
         private static readonly RepositoryTransactionFactory _createTransactionFactory
             = new RepositoryTransactionFactory();
-    }
-
-    internal static class InfractionQueryableExtensions
-    {
-        public static IQueryable<InfractionEntity> FilterInfractionsBy(this IQueryable<InfractionEntity> query, InfractionSearchCriteria criteria)
-        {
-            var longGuildId = (long?)criteria?.GuildId;
-            var longSubjectId = (long?)criteria?.SubjectId;
-            var longCreatedById = (long?)criteria?.CreatedById;
-
-            return query
-                .FilterBy(
-                    x => x.GuildId == longGuildId,
-                    longGuildId != null)
-                .FilterBy(
-                    x => criteria.Types.Contains(x.Type),
-                    criteria?.Types?.Any() ?? false)
-                .FilterBy(
-                    x => x.SubjectId == longSubjectId,
-                    longSubjectId != null)
-                .FilterBy(
-                    x => x.CreateAction.Created >= criteria.CreatedRange.Value.From,
-                    criteria?.CreatedRange?.From != null)
-                .FilterBy(
-                    x => x.CreateAction.Created <= criteria.CreatedRange.Value.To,
-                    criteria?.CreatedRange?.To != null)
-                .FilterBy(
-                    x => x.CreateAction.CreatedById == longCreatedById,
-                    longCreatedById != null)
-                .FilterBy(
-                    x => (x.RescindActionId != null) == criteria.IsRescinded,
-                    criteria?.IsRescinded != null)
-                .FilterBy(
-                    x => (x.DeleteActionId != null) == criteria.IsDeleted,
-                    criteria?.IsDeleted != null)
-                .FilterBy(
-                    x => (x.CreateAction.Created + x.Duration) >= criteria.ExpiresRange.Value.From,
-                    criteria?.ExpiresRange?.From != null)
-                .FilterBy(
-                    x => (x.CreateAction.Created + x.Duration) <= criteria.ExpiresRange.Value.To,
-                    criteria?.ExpiresRange?.To != null);
-        }
     }
 }
